TT-004 added setting functionalities [AM, EL]
