+++ app/build.gradle
@@ -59,7 +59,8 @@ dependencies {
     implementation "org.jetbrains.anko:anko:$anko_version"
     implementation "org.jetbrains.anko:anko-design:$anko_version"
     implementation "org.jetbrains.anko:anko-coroutines:$anko_version"
-
+    implementation 'androidx.wear:wear:1.0.0'
+    compileOnly 'com.google.android.wearable:wearable:2.6.0'
     def nav_version = "2.3.5"
 
     androidTestImplementation "androidx.navigation:navigation-testing:$nav_version"
