+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.TravelTogether">
-        <activity android:name=".SettingsActivity"></activity>
+        <activity android:name=".SettingActivity"/>
+        <activity
+            android:name=".ChangePassActivity"
+            android:label="@string/title_activity_change_pass"/>
         <activity android:name=".ProfileActivity" />
         <activity
             android:name=".StartActivity"
