+++ app/src/main/java/com/example/traveltogether/ChangePassActivity.kt
@@ -1,25 +1,29 @@
 package com.example.traveltogether
 
+import android.content.DialogInterface
+import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import android.widget.Button
 import android.widget.EditText
-import androidx.appcompat.app.AppCompatActivity
-import android.content.DialogInterface
-import android.content.Intent
 import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.app.AppCompatActivity
+import com.google.android.gms.tasks.OnCompleteListener
+import com.google.firebase.auth.EmailAuthProvider
+import com.google.firebase.auth.FirebaseAuth
 
 
 class ChangePassActivity : AppCompatActivity() {
     val positiveButtonClick = { dialog: DialogInterface, which: Int ->
         Toast.makeText(applicationContext,
                 android.R.string.yes, Toast.LENGTH_SHORT).show()
-        if (successfully)
+        if (successfully) {
             goBackToSetting();
+        }
     }
 
-    val alert : String = "Alert";
+    val alert : String = "Error";
     val success : String = "Success";
     var successfully : Boolean = false;
 
@@ -32,7 +36,6 @@ class ChangePassActivity : AppCompatActivity() {
         val new_pass: EditText = findViewById<View>(R.id.TextPassword_new) as EditText
         val new_pass_again: EditText = findViewById<View>(R.id.TextPassword_new_again) as EditText
 
-        print(old_pass.text);
         submit_btn.setOnClickListener {
             if (old_pass.text.toString().isEmpty())
                 PopUp(alert, "Please enter the old password");
@@ -46,11 +49,25 @@ class ChangePassActivity : AppCompatActivity() {
                 PopUp(alert,"Your old password and new one are the same, please change it!");
             else
             {
-                //TODO check if old password is same as in database
-                //TODO update the database with the new password!
-                successfully = true;
-                //after successfully changing it
-                PopUp(success, "You have changed your password successfully");
+                var cred = EmailAuthProvider.getCredential(FirebaseAuth.getInstance().currentUser.email,
+                    old_pass.text.toString()
+                );
+                var user =  FirebaseAuth.getInstance().currentUser;
+                if(user != null) {
+                    user.reauthenticate(cred)
+                        .addOnCompleteListener(OnCompleteListener<Void?> { task ->
+                            if (task.isSuccessful) {
+                                FirebaseAuth.getInstance().currentUser.updatePassword(new_pass.toString());
+                                successfully = true;
+                                PopUp(success, "You have changed your password successfully");
+                            } else {
+                                PopUp(alert, "old Password is wrong!");
+                            }
+                        })
+
+                }
+
+
             }
 
         }
