--- app/src/main/java/com/example/traveltogether/SettingsActivity.kt
@@ -1,27 +1,0 @@
-package com.example.traveltogether
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import org.jetbrains.anko.clearTask
-import org.jetbrains.anko.intentFor
-import org.jetbrains.anko.newTask
-
-class SettingsActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_settings)
-        val actionbar = supportActionBar
-        actionbar!!.title = "Settings"
-        actionbar.setDisplayHomeAsUpEnabled(true)
-        actionbar.setDisplayHomeAsUpEnabled(true)
-    }
-
-    override fun onSupportNavigateUp(): Boolean {
-        onBackPressed()
-        return true
-    }
-
-    override fun onBackPressed() {
-        startActivity(intentFor<MainActivity>().newTask().clearTask())
-    }
-}
