+++ app/src/main/res/values/strings.xml
@@ -27,7 +27,6 @@
     <string name="sign_in_text">Log In/Sign Up</string>
     <string name="sign_out_text">Sign Out</string>
     <string name="main_screen_text">MAIN SCREEN</string>
-    <string name="hello_world">Hello Square World!</string>
     <string name="profile_text">Profile</string>
     <string name="settings_text">Settings</string>
     <string name="logout_text">Logout</string>
